+++ CHANGELOG.rst
@@ -11,7 +11,8 @@ may want to subscribe to `GitHub's tag feed
 =====
 not released yet
 
-* new dependencies for setup: setuptools_scm
+* new dependency for running the tests: freezegun
+* new dependency for setup: setuptools_scm
 * khal now needs to be installed for building the documentation
 * khal should accept relative dates on the command line (today, tomorrow and 
   weekday names)